requirements.txt
bottle==0.11.6
distribute==0.6.34
python-mpd2==0.5.1
